Module aws_sdk_cognitoidentity::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

developer_user_already_registered_exception

See DeveloperUserAlreadyRegisteredException

external_service_exception

See ExternalServiceException

internal_error_exception

See InternalErrorException

invalid_identity_pool_configuration_exception

See InvalidIdentityPoolConfigurationException

invalid_parameter_exception

See InvalidParameterException

limit_exceeded_exception

See LimitExceededException

not_authorized_exception

See NotAuthorizedException

resource_conflict_exception

See ResourceConflictException

resource_not_found_exception

See ResourceNotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

ConcurrentModificationException

Thrown if there are parallel requests to modify a resource.

CreateIdentityPoolError

DeleteIdentitiesError

DeleteIdentityPoolError

DescribeIdentityError

DescribeIdentityPoolError

DeveloperUserAlreadyRegisteredException

The provided developer user identifier is already registered with Cognito under a

ExternalServiceException

An exception thrown when a dependent service such as Facebook or Twitter is not

GetCredentialsForIdentityError

GetIdError

GetIdentityPoolRolesError

GetOpenIdTokenError

GetOpenIdTokenForDeveloperIdentityError

GetPrincipalTagAttributeMapError

InternalErrorException

Thrown when the service encounters an error during processing the request.

InvalidIdentityPoolConfigurationException

Thrown if the identity pool has no role associated for the given auth type

InvalidParameterException

Thrown for missing or bad input parameter(s).

LimitExceededException

Thrown when the total number of user pools has exceeded a preset limit.

ListIdentitiesError

ListIdentityPoolsError

ListTagsForResourceError

LookupDeveloperIdentityError

MergeDeveloperIdentitiesError

NotAuthorizedException

Thrown when a user is not authorized to access the requested resource.

ResourceConflictException

Thrown when a user tries to use a login which is already linked to another

ResourceNotFoundException

Thrown when the requested resource (for example, a dataset or record) does not

SetIdentityPoolRolesError

SetPrincipalTagAttributeMapError

TagResourceError

TooManyRequestsException

Thrown when a request is throttled.

UnlinkDeveloperIdentityError

UnlinkIdentityError

UntagResourceError

UpdateIdentityPoolError

Enums

CreateIdentityPoolErrorKind

DeleteIdentitiesErrorKind

DeleteIdentityPoolErrorKind

DescribeIdentityErrorKind

DescribeIdentityPoolErrorKind

GetCredentialsForIdentityErrorKind

GetIdErrorKind

GetIdentityPoolRolesErrorKind

GetOpenIdTokenErrorKind

GetOpenIdTokenForDeveloperIdentityErrorKind

GetPrincipalTagAttributeMapErrorKind

ListIdentitiesErrorKind

ListIdentityPoolsErrorKind

ListTagsForResourceErrorKind

LookupDeveloperIdentityErrorKind

MergeDeveloperIdentitiesErrorKind

SetIdentityPoolRolesErrorKind

SetPrincipalTagAttributeMapErrorKind

TagResourceErrorKind

UnlinkDeveloperIdentityErrorKind

UnlinkIdentityErrorKind

UntagResourceErrorKind

UpdateIdentityPoolErrorKind