Module aws_sdk_cognitoidentity::model

Modules

cognito_identity_provider

See CognitoIdentityProvider

credentials

See Credentials

identity_description

See IdentityDescription

identity_pool_short_description

See IdentityPoolShortDescription

mapping_rule

See MappingRule

role_mapping

See RoleMapping

rules_configuration_type

See RulesConfigurationType

unprocessed_identity_id

See UnprocessedIdentityId

Structs

CognitoIdentityProvider

A provider representing an Amazon Cognito user pool and its client ID.

Credentials

Credentials for the provided identity ID.

IdentityDescription

A description of the identity.

IdentityPoolShortDescription

A description of the identity pool.

MappingRule

A rule that maps a claim name, a claim value, and a match type to a role

RoleMapping

A role mapping.

RulesConfigurationType

A container for rules.

UnprocessedIdentityId

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and

Enums

AmbiguousRoleResolutionType

ErrorCode

MappingRuleMatchType

RoleMappingType